airflow-proj-files/requirements.txt
apache-airflow-providers-papermill
Jupyter
matplotlib